requirements.txt
praw>=7.7.1
openai>=1.0.0
PyYAML==6.0.1
python-dotenv==1.0.0
streamlit>=1.28.0
pandas>=2.0.0